fix(studio): share safe media path cleanup
